data_simulator/requirements.txt
httpx
python-dotenv
PyYAML